+++ gpm_cache/api_helper.py
@@ -1,0 +1,6 @@
+class ApiHelper:
+    cached_library = None
+
+    @classmethod
+    def find_playlist(cls, api, playlist_name):
+        pass
