+++ setup.cfg
@@ -2,7 +2,7 @@
 test=pytest
 
 [tool:pytest]
-python_files = gpm_cache/tests.py
+python_files = tests/test*.py
 ; addopts =
 ;     --cov gpm_cache
 ;     --cov-report term-missing
