+++ tests/__init__.py
@@ -1,0 +1,4 @@
+import os
+
+REPO_ROOT = os.path.dirname(os.path.dirname(__file__))
+TEST_DATA_DIR = os.path.join(REPO_ROOT, 'tests', 'test_data')
